@@ -14,8 +14,8 @@ package main
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build wireinject
-// +build wireinject
+////go:build wireinject
+////+build wireinject
 
 // The build tag makes sure the stub is not built in the final build.
 
