@@ -5,23 +5,24 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
+	"time"
 
 	"github.com/hibiken/asynq"
 )
 
 // A list of task types.
 const (
-    TypeEmailDelivery   = "email:deliver"
-    TypeImageResize     = "image:resize"
+	TypeEmailDelivery = "email:deliver"
+	TypeImageResize   = "image:resize"
 )
 
 type EmailDeliveryPayload struct {
-    UserID     int
-    TemplateID string
+	UserID     int
+	TemplateID string
 }
 
 type ImageResizePayload struct {
-    SourceURL string
+	SourceURL string
 }
 
 //----------------------------------------------
@@ -30,20 +31,20 @@ type ImageResizePayload struct {
 //----------------------------------------------
 
 func NewEmailDeliveryTask(userID int, tmplID string) (*asynq.Task, error) {
-    payload, err := json.Marshal(EmailDeliveryPayload{UserID: userID, TemplateID: tmplID})
-    if err != nil {
-        return nil, err
-    }
-    return asynq.NewTask(TypeEmailDelivery, payload), nil
+	payload, err := json.Marshal(EmailDeliveryPayload{UserID: userID, TemplateID: tmplID})
+	if err != nil {
+		return nil, err
+	}
+	return asynq.NewTask(TypeEmailDelivery, payload), nil
 }
 
 func NewImageResizeTask(src string) (*asynq.Task, error) {
-    payload, err := json.Marshal(ImageResizePayload{SourceURL: src})
-    if err != nil {
-        return nil, err
-    }
-    // task options can be passed to NewTask, which can be overridden at enqueue time.
-    return asynq.NewTask(TypeImageResize, payload, asynq.MaxRetry(5), asynq.Timeout(20 * time.Minute)), nil
+	payload, err := json.Marshal(ImageResizePayload{SourceURL: src})
+	if err != nil {
+		return nil, err
+	}
+	// task options can be passed to NewTask, which can be overridden at enqueue time.
+	return asynq.NewTask(TypeImageResize, payload, asynq.MaxRetry(5), asynq.Timeout(20*time.Minute)), nil
 }
 
 //---------------------------------------------------------------
@@ -55,28 +56,28 @@ func NewImageResizeTask(src string) (*asynq.Task, error) {
 //---------------------------------------------------------------
 
 func HandleEmailDeliveryTask(ctx context.Context, t *asynq.Task) error {
-    var p EmailDeliveryPayload
-    if err := json.Unmarshal(t.Payload(), &p); err != nil {
-        return fmt.Errorf("json.Unmarshal failed: %v: %w", err, asynq.SkipRetry)
-    }
-    log.Printf("Sending Email to User: user_id=%d, template_id=%s", p.UserID, p.TemplateID)
-    // Email delivery code ...
-    return nil
+	var p EmailDeliveryPayload
+	if err := json.Unmarshal(t.Payload(), &p); err != nil {
+		return fmt.Errorf("json.Unmarshal failed: %v: %w", err, asynq.SkipRetry)
+	}
+	log.Printf("Sending Email to User: user_id=%d, template_id=%s", p.UserID, p.TemplateID)
+	// Email delivery code ...
+	return nil
 }
 
 // ImageProcessor implements asynq.Handler interface.
 type ImageProcessor struct {
-    // ... fields for struct
+	// ... fields for struct
 }
 
 func (processor *ImageProcessor) ProcessTask(ctx context.Context, t *asynq.Task) error {
-    var p ImageResizePayload
-    if err := json.Unmarshal(t.Payload(), &p); err != nil {
-        return fmt.Errorf("json.Unmarshal failed: %v: %w", err, asynq.SkipRetry)
-    }
-    log.Printf("Resizing image: src=%s", p.SourceURL)
-    // Image resizing code ...
-    return nil
+	var p ImageResizePayload
+	if err := json.Unmarshal(t.Payload(), &p); err != nil {
+		return fmt.Errorf("json.Unmarshal failed: %v: %w", err, asynq.SkipRetry)
+	}
+	log.Printf("Resizing image: src=%s", p.SourceURL)
+	// Image resizing code ...
+	return nil
 }
 
 func NewImageProcessor() *ImageProcessor {
