@@ -5,46 +5,49 @@ import (
 	"github.com/jmoiron/sqlx"
 )
 
-type Sqlx struct {
+type SqlxClient struct {
 	SqlxDB *sqlx.DB
 }
 
-func (s *Sqlx) Migrate(source string) error {
-	err := s.Open(source)
-	if err != nil {
-		return err
-	}
+func NewSqlxClient() *SqlxClient {
+	return &SqlxClient{}
+}
+
+func (s *SqlxClient) Migrate(source string) error {
 	m, err := migrate.NewWithDatabaseInstance(
-        "file:///migrations",
-        "postgres", nil)
-    return m.Up()
-}
-
-func (s *Sqlx) Open(source string) error {
-	db, err := sqlx.Connect("postgres", "user=foo dbname=bar sslmode=disable")
+		"file:///migrations",
+		"postgres", nil)
 	if err != nil {
 		return err
 	}
-	s.SqlxDB = db
-	return db.Ping()
+	return m.Up()
 }
 
-func (s *Sqlx) Close() error {
+//func (s *Sqlx) Open(source string) error {
+//	db, err := sqlx.Connect("postgres", "user=foo dbname=bar sslmode=disable")
+//	if err != nil {
+//		return err
+//	}
+//	s.SqlxDB = db
+//	return db.Ping()
+//}
+
+func (s *SqlxClient) Close() error {
 	//TODO implement me
 	panic("implement me")
 }
 
-func (s *Sqlx) Backend() {
+func (s *SqlxClient) Backend() {
 	//TODO implement me
 	panic("implement me")
 }
 
-func (s *Sqlx) Tx() error {
+func (s *SqlxClient) Tx() error {
 	//TODO implement me
 	panic("implement me")
 }
 
-func (s *Sqlx) Commit() error {
+func (s *SqlxClient) Commit() error {
 	//TODO implement me
 	panic("implement me")
 }
